requirements.txt
flask~=3.1.0
deepface~=0.0.93
tf-keras
tensorflow==2.12.0rc0
opencv-python-headless
numpy~=2.1.3
Pillow~=11.2.0

opencv-python~=4.11.0.86
flask-cors~=5.0.1